+++ sdk/src/relay/mod.rs
@@ -3,14 +3,14 @@
 use std::borrow::Cow;
 use std::cmp;
 use std::collections::HashMap;
-use std::future::Future;
 use std::sync::Arc;
 use std::time::Duration;
 
 use async_utility::time;
 use async_wsocket::ConnectionMode;
+use futures::StreamExt;
 use nostr_database::prelude::*;
-use tokio::sync::broadcast;
+use tokio::sync::{broadcast, oneshot};
 
 mod api;
 mod builder;
@@ -37,6 +37,7 @@ pub use self::stats::*;
 pub use self::status::*;
 use crate::client::ClientNotification;
 use crate::shared::SharedState;
+use crate::stream::{BoxedStream, NotificationStream};
 
 /// Subscription auto-closed reason
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -194,12 +195,40 @@ impl Relay {
         self.inner.set_notification_sender(notification_sender);
     }
 
-    /// Get new **relay** notification listener
+    /// Get a new notification stream
+    ///
+    /// The stream terminates when the relay shutdowns or is banned.
     ///
     /// <div class="warning">When you call this method, you subscribe to the notifications channel from that precise moment. Anything received by relay/s before that moment is not included in the channel!</div>
     #[inline]
-    pub fn notifications(&self) -> broadcast::Receiver<RelayNotification> {
-        self.inner.internal_notification_sender.subscribe()
+    pub fn notifications(&self) -> BoxedStream<RelayNotification> {
+        // If the relay is permanently unusable, return an empty stream
+        let status: RelayStatus = self.status();
+        if status.is_banned() || status.is_shutdown() {
+            return Box::pin(futures::stream::empty());
+        }
+
+        // Subscribe to notifications
+        let rx = self.inner.internal_notification_sender.subscribe();
+
+        // Create a oneshot channel
+        let (tx, rx_done) = oneshot::channel();
+        let mut tx: Option<oneshot::Sender<()>> = Some(tx);
+
+        Box::pin(
+            NotificationStream::new(rx)
+                .inspect(move |notification| {
+                    if let RelayNotification::RelayStatus { status } = &notification {
+                        if status.is_banned() || status.is_shutdown() {
+                            // Take the sender and send the oneshot notification
+                            if let Some(tx) = tx.take() {
+                                let _ = tx.send(());
+                            }
+                        }
+                    }
+                })
+                .take_until(rx_done),
+        )
     }
 
     /// Connect to the relay
@@ -410,28 +439,6 @@ impl Relay {
     #[inline]
     pub fn sync(&self, filter: Filter) -> SyncEvents {
         SyncEvents::new(self, filter)
-    }
-
-    /// Handle notifications
-    pub async fn handle_notifications<F, Fut>(&self, func: F) -> Result<(), Error>
-    where
-        F: Fn(RelayNotification) -> Fut,
-        Fut: Future<Output = Result<bool>>,
-    {
-        let mut notifications = self.notifications();
-        while let Ok(notification) = notifications.recv().await {
-            let shutdown: bool = match &notification {
-                RelayNotification::RelayStatus { status } => status.is_permanently_disconnected(),
-                _ => false,
-            };
-            let exit: bool = func(notification)
-                .await
-                .map_err(|e| Error::Handler(e.to_string()))?;
-            if exit || shutdown {
-                break;
-            }
-        }
-        Ok(())
     }
 }
 
@@ -1007,5 +1014,62 @@ mod tests {
 
         time::sleep(Duration::from_secs(5)).await;
         assert_eq!(relay.status(), RelayStatus::Connected);
+    }
+
+    #[tokio::test]
+    async fn test_terminate_notification_stream_on_shutdown() {
+        // Mock relay
+        let mock = MockRelay::run().await.unwrap();
+        let url = mock.url().await;
+
+        let relay = Relay::new(url);
+
+        // Connect
+        relay
+            .try_connect()
+            .timeout(Duration::from_secs(2))
+            .await
+            .unwrap();
+        assert_eq!(relay.status(), RelayStatus::Connected);
+
+        // Shutdown after some time
+        let r = relay.clone();
+        tokio::spawn(async move {
+            time::sleep(Duration::from_secs(3)).await;
+
+            r.shutdown()
+        });
+
+        let fut = async {
+            let mut notifications = relay.notifications();
+
+            let mut received = false;
+
+            // The stream must terminate after receiving the shutdown status
+            while let Some(n) = notifications.next().await {
+                if let RelayNotification::RelayStatus {
+                    status: RelayStatus::Shutdown,
+                } = n
+                {
+                    received = true;
+                }
+            }
+
+            // Make sure we received the shutdown status
+            assert!(received);
+        };
+        tokio::time::timeout(Duration::from_secs(5), fut)
+            .await
+            .unwrap();
+
+        // Try to get a new stream
+        let mut notifications = relay.notifications();
+
+        let res = tokio::time::timeout(Duration::from_secs(1), notifications.next())
+            .await
+            .unwrap();
+
+        // Must return None, as it's empty
+        assert!(res.is_none());
     }
 }