+++ sdk/src/relay/status.rs
@@ -127,11 +127,6 @@ impl RelayStatus {
     pub(crate) fn can_connect(&self) -> bool {
         matches!(self, Self::Initialized | Self::Terminated | Self::Sleeping)
     }
-
-    /// Check if relay can't reconnect again (status is `banned` or `shutdown`)
-    pub(super) fn is_permanently_disconnected(&self) -> bool {
-        matches!(self, Self::Banned | Self::Shutdown)
-    }
 }
 
 #[cfg(test)]
@@ -155,7 +150,6 @@ mod tests {
         assert!(!status.is_banned());
         assert!(!status.is_sleeping());
         assert!(!status.is_shutdown());
-        assert!(!status.is_permanently_disconnected());
         assert!(status.can_connect());
         let relay = AtomicRelayStatus::new(status);
         assert_eq!(relay.load(), RelayStatus::Initialized);
@@ -171,7 +165,6 @@ mod tests {
         assert!(!status.is_banned());
         assert!(!status.is_sleeping());
         assert!(!status.is_shutdown());
-        assert!(!status.is_permanently_disconnected());
         assert!(!status.can_connect());
         let relay = AtomicRelayStatus::new(status);
         assert_eq!(relay.load(), RelayStatus::Pending);
@@ -187,7 +180,6 @@ mod tests {
         assert!(!status.is_banned());
         assert!(!status.is_sleeping());
         assert!(!status.is_shutdown());
-        assert!(!status.is_permanently_disconnected());
         assert!(!status.can_connect());
         let relay = AtomicRelayStatus::new(status);
         assert_eq!(relay.load(), RelayStatus::Connecting);
@@ -203,7 +195,6 @@ mod tests {
         assert!(!status.is_banned());
         assert!(!status.is_sleeping());
         assert!(!status.is_shutdown());
-        assert!(!status.is_permanently_disconnected());
         assert!(!status.can_connect());
         let relay = AtomicRelayStatus::new(status);
         assert_eq!(relay.load(), RelayStatus::Connected);
@@ -219,7 +210,6 @@ mod tests {
         assert!(!status.is_banned());
         assert!(!status.is_sleeping());
         assert!(!status.is_shutdown());
-        assert!(!status.is_permanently_disconnected());
         assert!(!status.can_connect());
         let relay = AtomicRelayStatus::new(status);
         assert_eq!(relay.load(), RelayStatus::Disconnected);
@@ -235,7 +225,6 @@ mod tests {
         assert!(!status.is_banned());
         assert!(!status.is_sleeping());
         assert!(!status.is_shutdown());
-        assert!(!status.is_permanently_disconnected());
         assert!(status.can_connect());
         let relay = AtomicRelayStatus::new(status);
         assert_eq!(relay.load(), RelayStatus::Terminated);
@@ -251,7 +240,6 @@ mod tests {
         assert!(status.is_banned());
         assert!(!status.is_sleeping());
         assert!(!status.is_shutdown());
-        assert!(status.is_permanently_disconnected());
         assert!(!status.can_connect());
         let relay = AtomicRelayStatus::new(status);
         assert_eq!(relay.load(), RelayStatus::Banned);
@@ -267,7 +255,6 @@ mod tests {
         assert!(!status.is_banned());
         assert!(status.is_sleeping());
         assert!(!status.is_shutdown());
-        assert!(!status.is_permanently_disconnected());
         assert!(status.can_connect());
         let relay = AtomicRelayStatus::new(status);
         assert_eq!(relay.load(), RelayStatus::Sleeping);
@@ -283,7 +270,6 @@ mod tests {
         assert!(!status.is_banned());
         assert!(!status.is_sleeping());
         assert!(status.is_shutdown());
-        assert!(status.is_permanently_disconnected());
         assert!(!status.can_connect());
         let relay = AtomicRelayStatus::new(status);
         assert_eq!(relay.load(), RelayStatus::Shutdown);